+++ CHANGELOG.md
@@ -6,6 +6,10 @@ An EtherCAT master written in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- [#91] Add support for "cross" topologies, e.g. with EK1122.
+
 ## [0.2.1] - 2023-07-31
 
 ### Fixed
@@ -164,5 +168,6 @@ An EtherCAT master written in Rust.
 [#83]: https://github.com/ethercrab-rs/ethercrab/pull/83
 [#84]: https://github.com/ethercrab-rs/ethercrab/pull/84
 [#86]: https://github.com/ethercrab-rs/ethercrab/pull/86
+[#91]: https://github.com/ethercrab-rs/ethercrab/pull/91
 [0.2.0]: https://github.com/ethercrab-rs/ethercrab/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ethercrab-rs/ethercrab/compare/fb37346...v0.1.0